feat: created methods for focusing the cell by direction, move methods into separate files
